requirements.txt
python-dotenv
langchain
langchain-groq
langchain-community
chromadb